Add empty-fields-handling, improve robustness
Properties return `None` instead of returning 'N/A' for empty fields.

The code now checks if a variable is `None`-type before extracting
text out of it. Instead of using `.text.strip()`, it now uses
`.get_text(strip=True)`.

The `description` property no longer fails when the series has no
description. Before, it found the description using its `id`, but now it
fallsback to `entries['Description'] if no element with that `id` is
found.

`entries` chooses the first child of the found `<div>`s instead of the
entire text inside, so as to not include the text inside other elements
that are themselves inside `<div>`. For example, the `<div>` for
'Status' is (when prettified,):
```html
    <div class="sCat">
        <b>Status
            <div class="d-none d-sm-inline"> in Country of Origin</div>
        </b>
    </div>
```
In this case, `entries` chooses only 'Status' instead of 'Status in
Country of Origin', which is what it did previously. `entries` now also
strips its strings before using them as keys.

The order of if-statements in `latest_releases` was changed to
prioritize the checking of the strings that appear earlier.

The `href` optional argument of `find_all` is used instead of adding an
if-statement filtering out all elements without `href`.

Strings with parentheses appear often in the series page. The parens become
redundant as those strings are now separated from the other strings, so
`remove_outer_parens` was created.