Allow to export csv during iteration.
This makes it so that an error would not erase the progress done by the
function--especially important when it fetches ~100s of pages, which
previously would only be saved after it finished fetching.

Also fixed a bug on the while loop (replaced `<` with `<=`).

As the `csv` module is now used by both functions (and likely the next
ones as well), it's now imported globally.