Save to return value only if not exporting
If you are exporting the list, then you probably don't need to save the
data to be returned (note that this may reach megabytes of text). One
might see this as a function that is "doing too much". One way to avoid
this would be to do the exporting on another function, having inputs
that are the output of this function. This has the problem of taking a
long time (to the end of the job) before an output is produced, so if an
error occurs, all the previous work will be gone. This may be avoided by
using a `yield` instead of returning the entire list (TODO).

Allow to specify the first page to iterate on (useful for continuing
previous jobs). The function also accepts a `list_names` argument, if
only a specific list is to be used.

The csv writer now appends to the file (if it exists) instead of
overwriting.