Fix list_name bug, and more efficient resuming
The previous code assumed that there will always be only a single list
to be scraped, when resuming. This would have resulted to skipped
entries.

The previous code also checked the latest user id, so that it can resume
to the next user id of the same series. But since the writing to the
file occurs at the end of every list, in batches, the last entry will
always be the last one of the list of a series. Thus, checking that is
redundant. In this commit, that check is removed.