Use custom exceptions, allow them to occur
The main change here is that the code no longer tries to return some
default value if parsing failed. Instead it raises a parsing error. This
makes it easier to notice bugs. This also makes it so that the fields of
data classes are no longer required to always have a default value unless
it is truly optional.