Parse the color of recommendations
Mangaupdates' recommendations have a color associated with them, where a
darker shade means a better/closer recommendation. There is no
documentation for this, so the code simply sums the RGB values (instead
of averaging, as that would output floats, and is just scaled down by a
factor anyway, preserving relative values). A higher value is a "better"
recommendation (as opposed to mangaupdates where a darker shade = lower
intensity = "better" recommendation).