Change `read` method to `reading`, add new lists
To make the naming of the ListStats properties more consistent with the
naming scheme from MangaUpdates, the `read` property is now `reading`.
This also makes it less confusing (`read` could be interpreted as the
past tense, when it actually means "currently reading").

The `complete` and `on_hold` lists have been added.

Some of the comments have also been changed to be a bit cleaner.

The `general_list` method now uses `return` to signal that there are no
entries in the list, instead of `return None`. This makes it more
explicitly about stopping the iteration than actually returning a `None`
value.