Handle undecode-able characters better
The previously used HTML parser was 'lxml', which upon encountering an
undecodeable character, has unpredictable behavior. It has been replaced
by 'html.parser', which handles such cases better, by replacing the
character with '�'. The test for this bug has also been added.

Sidenote: This bug was found incidentally by parsing the top-N series
listed on MangaUpdates (see [1]). It probably has something to do with
the truncation done in the Description (which can be undone by clicking
'More...', which requires javascript to do).

[1] https://www.mangaupdates.com/stats.html?act=list&list=read